Introduce a deprecation framework for the policy DSL.
The primary focus of the framework is to be very simple to add
deprecations, and that those deprecations are shown when the
validate method is called directly using the CLI.

Deprecation checks are not run during the execution of the policies.

There is still some work to be done on the default formatting, and
on some of the duplciated context being shown.