A hunk of refactoring of the deprecations.
The deprecated filters and actions get the context during the
retrieval of the deprecations for a policy, so they don't need
to be initialized with the name. This provides much nicer reading
of the report.

Stopped trying to be clever by putting a single deprecation on
the same line as the heading. It didn't read very well smashed
together.

Each deprecation is now created with a unique ID. This will be
used for footnotes, which is to come in a following commit.
The footnotes will be shown at the end of the validation, after
all of the files and policies have been processed as a place
to show once extra information for each deprecation, like when
it will be removed, or become an error, and a possible link to
documentation.