Introduce a policy aware yaml loader.
The load_file utility function now constructs the PolicyLoader.
This loader has a FSM to determine the line number in the content
to be able to attach filename and line number information to the
metadata for the policy.

It was unclear in the docs if I'm correctly using the metadata
property of the policy. I couldn't find any reference to it other
than in the schema and structure validations of policies.